old-tests/submission/Microsoft/video/video_008.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>HTML5 Media Elements: 'media' attribute is 'all' by default.</title>
  <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
  <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
  <link rel="help" href="http://www.w3.org/TR/html5/video.html#the-source-element" />
  <meta name="assert" content="'media' attribute is 'all' by default." />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
  <script type="text/javascript" src="../common/media.js"></script>
  <script type="text/javascript">
            var videotest = async_test("'media' attribute is 'all' by default.");

function do_play(event)
{
    videotest.step(function() {
      var vid = document.getElementById("video0");
      assert_true(vid.currentSrc.indexOf("movie_300") > 0);
    });
    videotest.done();
}

</script>
</head>
<body>
<pre>
'media' attribute is 'all' by default.
</pre>
<p style="color: blue;">Wait until the text "Waiting ..." does not appear in the "Test Result" column below.</p>

<pre>

http://www.w3.org/TR/html5/video.html#the-source-element

The default, if the media attribute is omitted, is "all", meaning that 
by default the media resource is suitable for all media. 

</pre>
<div id='log'></div>

<video id="video0" autoplay onplay="do_play(event);">
    <script type="text/javascript">
    
        document.write(
            "<source src='" + getVideoURI("/media/movie_300") + "'" + 
            " />"
            );
        document.write(
            "<source src='" + getVideoURI("/media/movie_5") + "'" + 
            " media='all' />"
            );
    
    </script>
    Your browser does not support media elements.
</video>
</body>
</html>
